+++ commerce/delivery/src/main/java/ru/yandex/practicum/service/DeliveryService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -15,6 +16,6 @@ public interface DeliveryService {
 
     void deliveryPicked(UUID deliveryId);
 
-    Double deliveryCost(OrderDto dto);
+    BigDecimal deliveryCost(OrderDto dto);
 
 }
