+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/api/DeliveryApi.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryApi {
@@ -24,5 +25,5 @@ public interface DeliveryApi {
     void deliveryFailed(@RequestBody UUID deliveryId);
 
     @PostMapping("/api/v1/delivery/cost")
-    Double deliveryCost(@RequestBody @Valid OrderDto dto);
+    BigDecimal deliveryCost(@RequestBody @Valid OrderDto dto);
 }
